dl-7-3: OTA pull and dashboard stack, mode schedule, tests, BMad prompts
- OTAUpdater, GitHub release pull, dashboard/health UI and gzipped assets
- ModeOrchestrator/Registry scheduling, CustomLayoutMode V0 spike, TimeUtils
- Unity tests (OTA, schedule, modes), Playwright upgrade-banner e2e, smoke tests
- bmad-assist prompt archives; setup-skill-template tweaks; install-hooks.sh
- Untrack clangd index and Python bytecode; extend .gitignore

Made-with: Cursor

+++ firmware/interfaces/DisplayMode.h
@@ -16,7 +16,7 @@ struct RenderContext {
     uint16_t textColor;              // pre-computed from DisplayConfig RGB
     uint8_t brightness;              // read-only — managed by night mode scheduler
     uint16_t* logoBuffer;            // shared 2KB buffer from NeoMatrixDisplay
-    uint16_t displayCycleMs;         // cycle interval for modes that rotate flights
+    uint32_t displayCycleMs;         // cycle interval for modes that rotate flights (ms; uint32 avoids overflow for display_cycle > 65s)
 };
 
 // --- Zone Descriptor (static metadata for Mode Picker UI wireframes) ---
